+++ SinglySDK Example/SinglySDK Example/SinglyAppDelegate.m
@@ -35,6 +35,9 @@
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
 
+    // Initialize Defaults
+    [self initializeDefaults];
+
     // Initialize TestFlight
     #ifdef TESTFLIGHT_TOKEN
         [TestFlight takeOff:TESTFLIGHT_TOKEN];
@@ -51,6 +54,9 @@
     SinglySession *session = SinglySession.sharedSession;
     session.clientID = CLIENT_ID;
     session.clientSecret = CLIENT_SECRET;
+
+    // Set the Base URL (You don't need to do this in your own app)
+    session.baseURL = [[NSUserDefaults standardUserDefaults] objectForKey:@"SinglyApiLocation"];
 
     [[NSNotificationCenter defaultCenter] addObserverForName:kSinglySessionProfilesUpdatedNotification
                                                       object:self
@@ -74,6 +80,15 @@
          annotation:(id)annotation
 {
     return [SinglySession.sharedSession handleOpenURL:url];
+}
+
+#pragma mark - Defaults
+
+- (void)initializeDefaults
+{
+    NSString *defaultsFile = [[NSBundle mainBundle] pathForResource:@"Defaults" ofType:@"plist"];
+    NSDictionary *defaults = [NSDictionary dictionaryWithContentsOfFile:defaultsFile];
+    [[NSUserDefaults standardUserDefaults] registerDefaults:defaults];
 }
 
 @end